fix "/audio" for youtube handle links
If after the name of the channel there was some kind of addition through "/" (e.g. /audio or /video), then it is added at the end of the recognized channel token.
The link has been changed to https://www.youtube.com/@{username}/about, since "about" always weighs less than on the main page of the channel.
The "channel_handler_path" has been added, since an external module (podtube.py) is responsible for the request handler paths.